+++ tests/apps/test_app/management/commands/completion.py
@@ -1,12 +1,7 @@
 import json
-import sys
 import typing as t
 from pathlib import Path
-
-if sys.version_info < (3, 9):
-    from typing_extensions import Annotated
-else:
-    from typing import Annotated
+from typing import Annotated
 
 import typer
 from django.apps import AppConfig