+++ tests/apps/test_app/management/commands/error.py
@@ -1,9 +1,4 @@
-import sys
-
-if sys.version_info < (3, 9):
-    from typing_extensions import Annotated
-else:
-    from typing import Annotated
+from typing import Annotated
 
 from typer import Option