+++ tests/apps/test_app/management/commands/overloaded.py
@@ -1,10 +1,6 @@
 import json
 import sys
-
-if sys.version_info < (3, 9):
-    from typing_extensions import Annotated
-else:
-    from typing import Annotated
+from typing import Annotated
 
 from django.utils.translation import gettext_lazy as _
 from typer import Argument, Option